requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
